model/model.py
import torch
import torch.nn as nn
import torch.optim as optim
import numpy as np
from algorithm.harmonic_balancer import HarmonicBalancer
import spacy
from ethics.ethics_handler import EthicsHandler  # Import EthicsHandler


class Model:
    def __init__(self, input_dim, output_dim, archetype_name=None, context=None, _api_key=None, _service_url=None):
        self.input_dim = input_dim
        self.output_dim = output_dim
        self.archetype_name = archetype_name
        self.context = context
        self._api_key = _api_key
        self._service_url = _service_url
        self.harmonic_balancer = HarmonicBalancer(num_qubits=4, max_iterations=1000, harmony_memory_size=20)
        self.nlp_processor = NLPProcessor()
        self.ethics_handler = EthicsHandler()

        # Example neural network model
        self.model = nn.Sequential(
            nn.Linear(input_dim, 128),
            nn.ReLU(),
            nn.Linear(128, output_dim)
        )
        self.criterion = nn.MSELoss()
        self.optimizer = optim.Adam(self.model.parameters(), lr=0.001)

    def train_drl(self, data, targets):
        self.model.train()
        self.optimizer.zero_grad()
        outputs = self.model(data)
        loss = self.criterion(outputs, targets)
        loss.backward()
        self.optimizer.step()
        return loss.item()

    def train_upn(self, data, targets):
        self.model.train()
        self.optimizer.zero_grad()
        outputs = self.model(data)
        loss = self.criterion(outputs, targets)
        loss.backward()
        self.optimizer.step()
        return loss.item()

    def predict_drl(self, data):
        self.model.eval()
        with torch.no_grad():
            return self.model(data)

    def apply_ethical_principles(self, state):
        return self.ethics_handler.apply_principles(state)

    def optimize_parameters(self):
        print("Optimizing parameters using Harmonic Balancer...")
        best_solution, best_score = self.harmonic_balancer.run_experiment()
        # Update model parameters based on the best solution
        self.update_model_parameters(best_solution)

    def update_model_parameters(self, solution):
        # Logic to update model parameters based on the solution
        pass

    def save_model(self, path):
        torch.save(self.model.state_dict(), path)

    def load_model(self, path):
        self.model.load_state_dict(torch.load(path))
        self.model.eval()

class NLPProcessor:
    def __init__(self):
        self.nlp = spacy.load("en_core_web_sm")

    def process_message(self, message):
        doc = self.nlp(message)
        for entity in doc.ents:
            print(entity.text, entity.label_)

# Example usage
if __name__ == "__main__":
    model = Model(input_dim=10, output_dim=5)
    nlp_processor = NLPProcessor()
    nlp_processor.process_message("I want to teach children their heritage and spread the tradition of the old ways, along with the languages.")