tests/test_model.py
import unittest
from model.model import NLPProcessor
import torch
from model import Model
from ethics.ethics_handler import EthicsHandler
import spacy

class TestModel(unittest.TestCase):

    def setUp(self):
        self.input_dim = 10
        self.output_dim = 5
        self.model = Model(self.input_dim, self.output_dim)

    def test_model_initialization(self):
        self.assertIsNotNone(self.model)
        self.assertEqual(self.model.input_dim, self.input_dim)
        self.assertEqual(self.model.output_dim, self.output_dim)

    def test_train_drl(self):
        data = torch.randn(100, self.input_dim)
        targets = torch.randn(100, self.output_dim)
        loss = self.model.train_drl(data, targets)
        self.assertIsInstance(loss, float)

    def test_train_upn(self):
        data = torch.randn(100, self.input_dim)
        targets = torch.randn(100, self.output_dim)
        loss = self.model.train_upn(data, targets)
        self.assertIsInstance(loss, float)

    def test_predict_drl(self):
        data = torch.randn(10, self.input_dim)
        predictions = self.model.predict_drl(data)
        self.assertEqual(predictions.shape, (10, self.output_dim))

    def test_apply_ethical_principles(self):
        state = torch.randn(self.output_dim)
        enhanced_state = self.model.apply_ethical_principles(state)
        self.assertIsNotNone(enhanced_state)

class TestNLPProcessor(unittest.TestCase):

    def setUp(self):
        self.nlp_processor = NLPProcessor()

    def test_process_message(self):
        message = "I want to teach children their heritage and spread the tradition of the old ways, along with the languages."
        self.nlp_processor.process_message(message)
        # Add assertions based on expected output

if __name__ == "__main__":
    unittest.main()